requirements.txt
python-telegram-bot==21.6
requests==2.32.3
python-dotenv==1.0.0
SQLAlchemy==2.0.22
Flask==2.2.5
bcrypt==4.0.1
